island.is / timeline controls (#1125)
* rewire controls to native scroll and not transform

* darker disabled color

* add some documentation

* default value and typing

* current month

* condition for empty array

* Move timeline to island.is/web

* more specific interface naming

* fix disabled button sync

* remove timeline story
